Move "Update Flow Highlighting" button to visualisation layers
